Promote libsnes to an alternate UI.
If the Qt and Phoenix UIs are the SNES core linked against some extra
code that presents a friendly interface for users, libsnes is the SNES
core linked against some extra code that presents a friendly interface
for developers. Therefore, it seems only reasonable to make it an
alternate UI alongside the others.