Teach libsnes to automatically use -fPIC
My understanding is that -fPIC is not necessary for building shared
libraries on Intel 32-bit machines, but is required pretty much
everywhere else. Thus, I've taught nall to capture the target
architecture of the current compiler, and taught the libsnes makefile to
add -fPIC if it's not in a white-list of safe architectures.