Update the ldconfig cache when needed.
Previously we ran "ldconfig -n", which just updates the soname symlinks
in the given directory and doesn't update the central library cache.
This is needed so that other programs can dlopen("libsnes.so.1") without
supplying a path.

Also, we run ldconfig after uninstallation, so that it can clean up any
dangling symlinks.