Install libsnes with the full ABI version.
I read the Debian packaging guidelines, and it turns out that runtime
packages are expected to install "libfoo.so.1.2" (the full API version),
while ldconfig creates the soname symlink ("libfoo.so.1").
A hypothetical -devel package is expected to install the static library,
the "libfoo.so" symlink, and the header file.

Since the makefile installs both the runtime and development files,
we'll create all the symlinks for both packages ourselves.