Don't assume GLSL will implicitly convert values.
The GLSL 1.30 spec lists the allowed set of implicit conversions in
section 4.1.10, and none of those conversion involve changing a vector
from one size to another. Nevertheless, the original Pixellate shader
tried to grab the basic texture coordinates out of gl_TexCoord[0] (a
vec4) by using it in a context that required a vec2. Explicitly
extracting the texture coordinates as a vec2 by appending ".st" makes
things more portable.