Avoid division-by-zero when calculating coeffs.
The original code tried to avoid division-by-zero with a complicated
mix() filter; a much simpler system is just to add a very small value
each field of the vector. That way we can remove a complex function
call, the custom less_than() function, and restore the obvious
calculation for the "one" value.