Don't let libsnes export internal-only symbols.
libsnes is implemented with a variety of technologies including libco
and nall, and could potentially be loaded into a host application that
uses its own, conflicting functions with the same names (especially
nall, which gets rewritten occasionally). For safety, use a linker
version-script to mark all the official snes_* functions as external,
and everything else as internal.