Teach the build system to make portable builds.
higan and icarus support looking for their support files in the same
directory as the executable, which means you can download an archive
with a pre-built binary and run it without having to install it. To make
sure these archives are consistently built the same way, let's put the
steps into the Makefile.