Teach nall to seed from /dev/urandom if it can.
The getrandom() syscall is still fairly new on Linux, and is mostly useful
for processes that (a) run early in the boot process, before much entropy
is collected, or (b) are worried about file-descriptor exhaustion. higan is
neither of those, and Wikipedia says /dev/urandom is present on Linux, the BSDs,
Solaris, AIX, HP-UX and Tru64 UNIX, so it's probably pretty safe. If we can't
open that device for whatever reason, we can fall back to seeding from the
system time as before.